requirement.txt
torch>=1.6.0
torchaudio>=0.6.0
numpy
scipy
scikit-learn
tqdm
scenedetect
opencv-python
python_speech_features
torchvision
ffmpeg
gdown
youtube-dl
pandas
matplotlib
pillow